Bring utility functions common to both Client and CLS in a single utils module
* Remove `cls_arrow_utils.cc` from CLS

* Fix some headers
