+++ cpp/src/arrow/dataset/api.h
@@ -28,3 +28,4 @@
 #include "arrow/dataset/file_parquet.h"
 #include "arrow/dataset/filter.h"
 #include "arrow/dataset/scanner.h"
+#include "arrow/dataset/rados_utils.h"
